chore: move TokenDecoder to its own file
