+++ src/utils/helpers/index.js
@@ -1,7 +1,9 @@
 const Encrypter = require('./encrypter');
+const TokenDecoder = require('./token-decoder');
 const TokenGenerator = require('./token-generator');
 
 module.exports = {
   Encrypter,
+  TokenDecoder,
   TokenGenerator
 };
